FIX: Use correct GLint type (rather than long) in ApplePBTextureTarget to resolve potential compilation issues if GLint is typedef'd to an incompatible type.  See: http://www.cegui.org.uk/phpBB2/viewtopic.php?f=3&t=4553

+++ cegui/include/RendererModules/OpenGL/CEGUIOpenGLApplePBTextureTarget.h
@@ -76,7 +76,7 @@ protected:
     //! The GL context that's used by this TextureTarget.
     CGLContextObj d_context;
     //! virtual screen used by the "parent" context
-    long d_screen;
+    GLint d_screen;
     //! Context that was active before ours was activated.
     mutable CGLContextObj d_prevContext;
 };
